guard npc working with animations

+++ source/shaders/skinning.vert
@@ -6,7 +6,7 @@ layout(location = 0) in vec3 inPosition;
 layout(location = 1) in vec2 inUV;
 layout(location = 2) in vec3 inNormal;
 layout(location = 3) in vec4 inWeights;
-layout(location = 4) in vec4 inJoints;
+layout(location = 4) in uvec4 inJoints;
 
 // --- UNIFORM BUFFERS ---
 // Set 1, Binding 0: Il tuo AnimUniformBufferObject
